pyproject.toml
[tool.poetry]
package-mode = false
name = "django-mai"
version = "0.1.0"
description = "Django DRF проект с аутентификацией"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.14"
python-decouple = "^3.8"  # Для работы с .env файлами
psycopg2-binary = "^2.9"  # PostgreSQL драйвер
django-cors-headers = "^4.3"  # Для CORS
gunicorn = "^21.2"  # WSGI сервер для продакшена
drf-yasg = "^1.21.10"
pillow = "^11.2.1"
djangorestframework-simplejwt = "^5.5.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"

[tool.ruff]
exclude = [
    ".git",
    ".git-rewrite",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pycache__",
    "__pypackages__",
]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "D",
    "UP",
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "PIE",
    "T20",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "ERA",
    "PGH",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "TRY",
    "FLY",
    "PERF",
    "RUF",
    "D213",
]
ignore = [
    "D101",
    "D104",
    "D107",
    "D100",
    "D212",
    "D300",
    "COM812",
    "T201",
    "RUF002",
    "ERA001",
    "INP001",
    "RUF012",
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.TypedDict".msg = "Use typing_extensions.TypedDict instead."

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
docstring-quotes = "single"
multiline-quotes = "single"

[tool.ruff.lint.isort]
force-to-top = ["typing"]
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["S101", "S104", "S106"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
docstring-code-format = false
docstring-code-line-length = "dynamic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 